+++ core/include/drivers/stm32_gpio.h
@@ -7,68 +7,13 @@
 #define DRIVERS_STM32_GPIO_H
 
 #include <assert.h>
-#include <compiler.h>
 #include <drivers/pinctrl.h>
 #include <stdbool.h>
 #include <stdint.h>
 #include <stddef.h>
 
-#define GPIO_MODE_INPUT		0x0
-#define GPIO_MODE_OUTPUT	0x1
-#define GPIO_MODE_ALTERNATE	0x2
-#define GPIO_MODE_ANALOG	0x3
-
-#define GPIO_OTYPE_PUSH_PULL	0x0
-#define GPIO_OTYPE_OPEN_DRAIN	0x1
-
-#define GPIO_OSPEED_LOW		0x0
-#define GPIO_OSPEED_MEDIUM	0x1
-#define GPIO_OSPEED_HIGH	0x2
-#define GPIO_OSPEED_VERY_HIGH	0x3
-
-#define GPIO_PUPD_NO_PULL	0x0
-#define GPIO_PUPD_PULL_UP	0x1
-#define GPIO_PUPD_PULL_DOWN	0x2
-
-#define GPIO_OD_LEVEL_LOW	0x0
-#define GPIO_OD_LEVEL_HIGH	0x1
-
 struct pinctrl_state;
-
-/*
- * GPIO configuration description structured as single 16bit word
- * for efficient save/restore when GPIO pin suspends or resumes.
- *
- * @mode: One of GPIO_MODE_*
- * @otype: One of GPIO_OTYPE_*
- * @ospeed: One of GPIO_OSPEED_*
- * @pupd: One of GPIO_PUPD_*
- * @od: One of GPIO_OD_*
- * @af: Alternate function numerical ID between 0 and 15
- */
-struct gpio_cfg {
-	uint16_t mode:		2;
-	uint16_t otype:		1;
-	uint16_t ospeed:	2;
-	uint16_t pupd:		2;
-	uint16_t od:		1;
-	uint16_t af:		4;
-};
-
-/*
- * Descrption of a pin and its 2 states muxing
- *
- * @bank: GPIO bank identifier as assigned by the platform
- * @pin: Pin number in the GPIO bank
- * @cfg: Pin configuration
- * @active_cfg: Configuration in active state
- * @standby_cfg: Configuration in standby state
- */
-struct stm32_pinctrl {
-	uint8_t bank;
-	uint8_t pin;
-	struct gpio_cfg cfg;
-};
+struct stm32_pinctrl;
 
 /*
  * Save pinctrl instances defined in DT node: identifiers and power states
